Increase adaptive hydro grid size.
Determine rmax at 110 MeV on a 27 fm grid.  Testing revealed that the
old parameters (120 MeV and 25 fm) were a little too tight.  With the
new parameters, there is no difference in computed observables between the
adaptive grid and a fixed 27 fm grid, but the adaptive grid is roughly
twice as fast on average.